src/agegate-bot/requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
cryptography>=41.0.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
pytesseract>=0.3.10
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
itsdangerous>=2.1.0
